// backend/src/types/review.d.ts
interface Review {
  id: number;
  eventId: number;
  userId: number;
  reviewText: string;
  imageUrl1: string;
  imageUrl2: string;
  imageUrl3: string;
  imageUrl4: string;
  imagePublicId1: string;
  imagePublicId2: string;
  imagePublicId3: string;
  imagePublicId4: string;
}

export default Review;